dbnet: fix the failure in polygon mode
Fixing the failure listed below in polygon mode

Traceback (most recent call last):                                                                                                                                        │·············
  File "dbnet-filter.py", line 174, in <module>                                                                                                                           │·············
    start_main()                                                                                                                                                          │·············
  File "dbnet-filter.py", line 166, in start_main                                                                                                                         │·············
    poly=args.poly, viz=args.viz)                                                                                                                                         │·············
  File "/mnt/DATA3/devesh/projects/datageneration/dbnet/dbnet.py", line 31, in run_dbnet                                                                                  │·············
    return DBN(experiment, experiment_args, cmd=args).inference(args['image_path'], args['visualize'])                                                                    │·············
  File "/mnt/DATA3/devesh/projects/datageneration/dbnet/dbnet.py", line 127, in inference                                                                                 │·············
    is_output_polygon=self.args['polygon'])                                                                                                                               │·············
  File "/mnt/DATA3/devesh/projects/datageneration/dbnet/structure/representers/seg_detector_representer.py", line 53, in represent                                        │·············
    segmentation[batch_index], width, height)                                                                                                                             │·············
  File "/mnt/DATA3/devesh/projects/datageneration/dbnet/structure/representers/seg_detector_representer.py", line 102, in polygons_from_bitmap                            │·············
    _, sside = self.get_mini_boxes(box.reshape((-1, 1, 2)))                                                                                                               │·············
  File "/mnt/DATA3/devesh/projects/datageneration/dbnet/structure/representers/seg_detector_representer.py", line 171, in get_mini_boxes                                  │·············
    bounding_box = cv2.minAreaRect(contour)                                                                                                                               │·············
cv2.error: OpenCV(4.1.2) /io/opencv/modules/imgproc/src/convhull.cpp:137: error: (-215:Assertion failed) total >= 0 && (depth == CV_32F || depth == CV_32S) in function 'c│·············
onvexHull'

+++ structure/representers/seg_detector_representer.py
@@ -98,7 +98,7 @@ class SegDetectorRepresenter(Configurable):
                     continue
             else:
                 continue
-            box = box.reshape(-1, 2)
+            box = box.reshape(-1, 2).astype(np.int64)
             _, sside = self.get_mini_boxes(box.reshape((-1, 1, 2)))
             if sside < self.min_size + 2:
                 continue
